SLM-62 Use modules for prisonTypes
