+++ server/routes/barcode/contacts/ChooseContactController.ts
@@ -1,7 +1,7 @@
 import { Request, Response } from 'express'
 import moment from 'moment'
+import type { PrisonAddress, Recipient } from 'prisonTypes'
 import PrisonRegisterService from '../../../services/prison/PrisonRegisterService'
-import { PrisonAddress, Recipient } from '../../../@types/prisonTypes'
 import ContactService from '../../../services/contacts/ContactService'
 import type { Contact } from '../../../@types/sendLegalMailApiClientTypes'
 import ChooseContactView from './ChooseContactView'
