+++ server/routes/barcode/contacts/CreateContactByPrisonerNameController.ts
@@ -1,8 +1,8 @@
 import { Request, Response } from 'express'
 import type { CreateNewContactByPrisonerNameForm } from 'forms'
 import moment from 'moment'
+import type { Prison, PrisonAddress, Recipient } from 'prisonTypes'
 import PrisonRegisterService from '../../../services/prison/PrisonRegisterService'
-import { Prison, PrisonAddress, Recipient } from '../../../@types/prisonTypes'
 import CreateContactByPrisonerNameView from './CreateContactByPrisonerNameView'
 import validateNewContact from './newContactByPrisonerNameValidator'
 import filterSupportedPrisons from './filterSupportedPrisons'
