+++ server/routes/barcode/contacts/CreateContactByPrisonerNameView.ts
@@ -1,5 +1,5 @@
 import type { CreateNewContactByPrisonerNameForm } from 'forms'
-import { Prison } from '../../../@types/prisonTypes'
+import type { Prison } from 'prisonTypes'
 import getPrisonDropdown, { DropDownOption } from './prisonDropdown'
 
 export default class CreateContactByPrisonerNameView {
